Refactor PageHeaderImage
