+++ packages/ui/src/PageHeader/PageHeaderImage.tsx
@@ -1,30 +1,17 @@
 import { clsx } from 'clsx'
-import Image, { type ImageProps } from 'next/image'
+import Image from 'next/image'
 
-import { buildImageSizeProp } from '@filecoin-foundation/utils/buildImageSizeProp'
-import type {
-  ImageObjectFit,
-  StaticImageProps,
-} from '@filecoin-foundation/utils/types/imageType'
-
-export type PageHeaderImageProps = (StaticImageProps | ImageProps) & {
-  objectFit?: ImageObjectFit
-}
+import {
+  getPageHeaderImageProps,
+  type PageHeaderImageProps,
+} from './pageHeaderImageHelpers'
 
 export function PageHeaderImage(image: PageHeaderImageProps) {
   const isStaticImage = 'data' in image
-
-  const commonProps = {
-    alt: image.alt,
-    priority: true,
-    quality: 100,
-    sizes: buildImageSizeProp({ startSize: '100vw', lg: '490px' }),
-    className: clsx(
-      'page-header-image',
-      image.objectFit === 'cover' && 'object-cover',
-      image.objectFit === 'contain' && 'object-contain',
-    ),
-  }
+  const commonProps = getPageHeaderImageProps(image, {
+    startSize: '100vw',
+    lg: '490px',
+  })
 
   if (isStaticImage) {
     return (
